+++ android/README.md
@@ -1,4 +1,4 @@
-# StravaAlt Android
+# Travalt Android
 
 This module will contain the future Android application written in Kotlin.
 Currently it's a minimal placeholder project.
