+++ backend/src/main/resources/application.conf
@@ -3,6 +3,6 @@ ktor {
         port = 8080
     }
     application {
-        modules = [ com.stravaalt.ApplicationKt.module ]
+        modules = [ com.travalt.ApplicationKt.module ]
     }
 }
