+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "strava-alt-frontend",
+  "name": "travalt-frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
